change subscript predicate index type to size_t since it must be positive

+++ src/main/c/include/jsonpath.h
@@ -58,7 +58,7 @@ struct predicate
     {
         struct
         {
-            uint_fast32_t index;
+            size_t index;
         } subscript;
         
         struct
@@ -187,7 +187,7 @@ bool       step_has_predicate(const step * restrict step);
 predicate *step_get_predicate(const step * restrict step);
 
 enum predicate_kind predicate_get_kind(const predicate * restrict predicate);
-uint_fast32_t       subscript_predicate_get_index(const predicate * restrict predicate);
+size_t              subscript_predicate_get_index(const predicate * restrict predicate);
 uint_fast32_t       slice_predicate_get_to(const predicate * restrict predicate);
 uint_fast32_t       slice_predicate_get_from(const predicate * restrict predicate);
 uint_fast32_t       slice_predicate_get_step(const predicate * restrict predicate);
